tsd/scripting: Split Lua bindings TUs to parallelise compile
`tsd_lua` build wall dropped 74s → 17s by splitting sol2-heavy
translation units that previously serialised the whole build:

  - ObjectBindings.cpp (was 72s): nine ObjectPoolRef<T>
    registrations moved to bindings/refs/{Type}Ref.cpp; method bodies
    pulled into non-template free functions in ObjectMethodImpls.cpp;
    array helpers extracted to ArrayHelpers.cpp.
  - CoreBindings.cpp (was 31s): Object usertype moved to
    ObjectUsertype.cpp; Animation/AnimationManager moved to
    AnimationBindings.cpp; Scene's ~46 methods split across five
    bindings/scene/{Accessors,Creators,Iteration,Layers,Nodes}.cpp.
  - MathBindings.cpp (was 13s): five usertypes split into
    bindings/math/{Float2,Float3,Float4,Mat3,Mat4}.cpp with a shared
    RegisterVec.hpp for vec arithmetic.
