+++ tsd/src/tsd/scripting/bindings/ObjectUsertype.hpp
@@ -1,0 +1,12 @@
+// Copyright 2026 NVIDIA Corporation
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <sol/sol.hpp>
+
+namespace tsd::scripting {
+
+void registerObjectUsertype(sol::table &tsd);
+
+} // namespace tsd::scripting
